task_me_api/requirements.txt
firebase-admin==5.2.0
Flask==2.1.1
PyJWT==2.3.0
google-cloud-firestore